+++ examples/01_basic_bringup_cli/main.cpp
@@ -1,6 +1,10 @@
+#if defined(AT21CS_EXAMPLE_PLATFORM_IDF)
+#include "../common/IdfArduinoCompat.h"
+#else
 #include <Arduino.h>
+#endif
 
-#if defined(ARDUINO_ARCH_ESP32)
+#if defined(ARDUINO_ARCH_ESP32) || defined(AT21CS_EXAMPLE_PLATFORM_IDF)
 #include <driver/gpio.h>
 #include <esp_timer.h>
 #if __has_include(<esp_cpu.h>)