+++ examples/common/BoardConfig.h
@@ -5,7 +5,11 @@
 
 #pragma once
 
+#if defined(AT21CS_EXAMPLE_PLATFORM_IDF)
+#include "IdfArduinoCompat.h"
+#else
 #include <Arduino.h>
+#endif
 
 namespace board {