+++ examples/common/BusDiag.h
@@ -1,6 +1,10 @@
 #pragma once
 
+#if defined(AT21CS_EXAMPLE_PLATFORM_IDF)
+#include "IdfArduinoCompat.h"
+#else
 #include <Arduino.h>
+#endif
 
 namespace bus_diag {
 inline void scan() {