+++ examples/common/Log.h
@@ -8,7 +8,11 @@
 
 #pragma once
 
+#if defined(AT21CS_EXAMPLE_PLATFORM_IDF)
+#include "IdfArduinoCompat.h"
+#else
 #include <Arduino.h>
+#endif
 
 #include "examples/common/BuildConfig.h"